+++ dbus_remote_proxy/lib/dbus_remote_signal_stream_proxy.dart
@@ -75,7 +75,7 @@ class DBusRemoteObjectSignalStreamProxy extends DBusRemoteObjectSignalStream {
 
     final request = jsonEncode({
       'serviceName': object.name,
-      'serviceType': object.type,
+      'serviceType': object.clientProxy.serviceType,
       'interface': interface,
       'member': name,
       'path': object.path.value,